refactor: Kafka 알림 어댑터 패키지 정리
