+++ nolzo-api/src/test/java/com/noljo/nolzo/notification/adapter/kafka/KafkaSeatAvailableEventPublisherTest.java
@@ -1,4 +1,4 @@
-package com.noljo.nolzo.notification.adapter.out.event;
+package com.noljo.nolzo.notification.adapter.kafka;
 
 import static org.mockito.Mockito.verify;
