+++ nolzo-api/src/test/resources/application-test.yml
@@ -13,7 +13,7 @@ spring:
       key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
       value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
       properties:
-        spring.json.trusted.packages: com.noljo.nolzo.notification.adapter.out.event
+        spring.json.trusted.packages: com.noljo.nolzo.notification.adapter.kafka
     listener:
       auto-startup: false
   data:
