+++ nolzo-infra/src/main/java/com/noljo/nolzo/notification/adapter/kafka/KafkaSeatAvailableEventPublisher.java
@@ -1,4 +1,4 @@
-package com.noljo.nolzo.notification.adapter.out.event;
+package com.noljo.nolzo.notification.adapter.kafka;
 
 import com.noljo.nolzo.notification.application.port.out.PublishSeatAvailableEventPort;
 import com.noljo.nolzo.notification.domain.event.SeatAvailableEvent;
