+++ nolzo-infra/src/main/java/com/noljo/nolzo/notification/adapter/kafka/SeatAvailableEventConsumer.java
@@ -1,6 +1,5 @@
-package com.noljo.nolzo.notification.adapter.in.event;
+package com.noljo.nolzo.notification.adapter.kafka;
 
-import com.noljo.nolzo.notification.adapter.out.event.SeatAvailableEventMessage;
 import com.noljo.nolzo.notification.application.port.in.HandleSeatAvailableUseCase;
 import com.noljo.nolzo.notification.domain.event.SeatAvailableEvent;
 import lombok.RequiredArgsConstructor;
